Adds patient profile edit functionality
Adds the ability to edit patient profiles, including therapies, allergies, and studies.

This change introduces a new view for editing patient profiles, linked from the patient detail page. It also updates the forms and URLs to support this functionality.
The therapy update view was removed, and therapies are now edited directly via the patient edit page.

The success URL for ResearchAnalysisUpdateView was updated to redirect to patient detail view.